Split player view into separate files
